split manager from module params
