+++ params/alias.go
@@ -12,19 +12,7 @@ const (
 
 var (
 	// functions aliases
-	RegisterCodec              = types.RegisterCodec
-	ErrUnknownSubspace         = types.ErrUnknownSubspace
-	ErrSettingParameter        = types.ErrSettingParameter
-	NewParameterChangeProposal = types.NewParameterChangeProposal
-	NewParamChange             = types.NewParamChange
-	NewCodec                   = types.NewCodec
-
-	// variable aliases
-	ModuleCdc = types.ModuleCdc
+	ErrUnknownSubspace  = types.ErrUnknownSubspace
+	ErrSettingParameter = types.ErrSettingParameter
 )
 
-type (
-	Codec                   = types.Codec
-	ParameterChangeProposal = types.ParameterChangeProposal
-	ParamChange             = types.ParamChange
-)
