+++ params/types/subspace/subspace_test.go
@@ -11,10 +11,10 @@ import (
 	dbm "github.com/tendermint/tm-db"
 
 	"github.com/cosmos/cosmos-sdk/codec"
-	ptypes "github.com/cosmos/cosmos-sdk/params"
 	"github.com/cosmos/cosmos-sdk/params/types/subspace"
 	"github.com/cosmos/cosmos-sdk/store"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	ptypes "github.com/cosmos/cosmos-sdk/x/params"
 )
 
 type SubspaceTestSuite struct {
