+++ x/params/internal/types/proposal_test.go
@@ -4,6 +4,8 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/require"
+
+	"github.com/cosmos/cosmos-sdk/params/types"
 )
 
 func TestParameterChangeProposal(t *testing.T) {
@@ -13,7 +15,7 @@ func TestParameterChangeProposal(t *testing.T) {
 
 	require.Equal(t, "test title", pcp.GetTitle())
 	require.Equal(t, "test description", pcp.GetDescription())
-	require.Equal(t, RouterKey, pcp.ProposalRoute())
+	require.Equal(t, types.RouterKey, pcp.ProposalRoute())
 	require.Equal(t, ProposalTypeChange, pcp.ProposalType())
 	require.Nil(t, pcp.ValidateBasic())
