+++ x/params/proposal_handler.go
@@ -2,6 +2,7 @@ package params
 
 import (
 	"fmt"
+
 	"github.com/cosmos/cosmos-sdk/params"
 
 	"github.com/cosmos/cosmos-sdk/params/manager"
@@ -14,7 +15,7 @@ import (
 func NewParamChangeProposalHandler(k manager.Manager) govtypes.Handler {
 	return func(ctx sdk.Context, content govtypes.Content) error {
 		switch c := content.(type) {
-		case params.ParameterChangeProposal:
+		case ParameterChangeProposal:
 			return handleParameterChangeProposal(ctx, k, c)
 
 		default:
@@ -23,7 +24,7 @@ func NewParamChangeProposalHandler(k manager.Manager) govtypes.Handler {
 	}
 }
 
-func handleParameterChangeProposal(ctx sdk.Context, k manager.Manager, p params.ParameterChangeProposal) error {
+func handleParameterChangeProposal(ctx sdk.Context, k manager.Manager, p ParameterChangeProposal) error {
 	for _, c := range p.Changes {
 		ss, ok := k.GetSubspace(c.Subspace)
 		if !ok {
