+++ x/params/proposal_handler_test.go
@@ -1,8 +1,9 @@
 package params_test
 
 import (
-	params2 "github.com/cosmos/cosmos-sdk/x/params"
 	"testing"
+
+	params2 "github.com/cosmos/cosmos-sdk/x/params"
 
 	"github.com/stretchr/testify/require"
 	abci "github.com/tendermint/tendermint/abci/types"
@@ -10,11 +11,11 @@ import (
 	dbm "github.com/tendermint/tm-db"
 
 	"github.com/cosmos/cosmos-sdk/codec"
-	"github.com/cosmos/cosmos-sdk/params"
 	"github.com/cosmos/cosmos-sdk/params/manager"
 	"github.com/cosmos/cosmos-sdk/params/types/subspace"
 	"github.com/cosmos/cosmos-sdk/store"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/x/params"
 )
 
 func validateNoOp(_ interface{}) error { return nil }
